extended and fixed kafka features

+++ server/kafka/protocol/offsetFetch/offsetFetch.go
@@ -46,7 +46,7 @@ type ResponseTopic struct {
 type Partition struct {
 	Index                int32            `kafka:""`
 	CommittedOffset      int64            `kafka:""`
-	CommittedLeaderEpoch int64            `kafka:"min=5"`
+	CommittedLeaderEpoch int32            `kafka:"min=5"`
 	Metadata             string           `kafka:"compact=6,nullable"`
 	ErrorCode            int16            `kafka:""`
 	TagFields            map[int64]string `kafka:"type=TAG_BUFFER,min=6"`
